+++ tests/testthat/test-geo_address_lookup_sf.R
@@ -1,19 +1,23 @@
 test_that("Returning empty query", {
-  obj <- expect_message(geo_address_lookup_sf("xbzbzbzoa aiaia", "R"))
-
-  expect_true(ncol(obj) == 1)
+  expect_message(geo_address_lookup_sf("xbzbzbzoa aiaia", "R"))
+  skip_if_api_server()
+  
+  obj <- geo_address_lookup_sf("xbzbzbzoa aiaia", "R")
+  
   expect_true(nrow(obj) == 1)
 
   expect_true(obj$query == "Rxbzbzbzoa aiaia")
 })
 
 test_that("Data format", {
+  skip_if_api_server()
   skip_if_offline()
   expect_true(is.data.frame(geo_address_lookup_sf(34633854, "W")))
   expect_s3_class(geo_address_lookup_sf(34633854, "W"), "sf")
 })
 
 test_that("Checking query", {
+  skip_if_api_server()
   skip_if_offline()
   expect_equal(ncol(geo_address_lookup_sf(34633854, "W")), 3)
   expect_gt(ncol(geo_address_lookup_sf(34633854, "W", full_results = TRUE)), 3)