+++ tests/testthat/test-geo_amenity.R
@@ -1,19 +1,26 @@
 test_that("Returning empty query", {
-  obj <- expect_message(geo_amenity(
+  expect_message(geo_amenity(
     bbox = c(-1.1446, 41.5022, -0.4854, 41.8795),
     amenity = "xbzbzbzoa aiaia"
   ))
+  
+  skip_if_api_server()
+  
+  obj <- geo_amenity_sf(
+    bbox = c(-1.1446, 41.5022, -0.4854, 41.8795),
+    amenity = "xbzbzbzoa aiaia"
+  )
 
-  expect_true(ncol(obj) == 3)
   expect_true(nrow(obj) == 1)
 
   expect_true(obj$query == "xbzbzbzoa aiaia")
-  expect_true(is.na(obj$lat))
-  expect_true(is.na(obj$lon))
+  
 })
 
 test_that("Data format", {
+  skip_if_api_server()
   skip_if_offline()
+
   expect_true(is.data.frame(geo_amenity(
     bbox = c(-1.1446, 41.5022, -0.4854, 41.8795),
     c("pub", "restaurant"),
@@ -25,7 +32,9 @@ test_that("Data format", {
 })
 
 test_that("Checking query", {
+  skip_if_api_server()
   skip_if_offline()
+
   expect_equal(ncol(geo_amenity(
     bbox = c(-1.1446, 41.5022, -0.4854, 41.8795),
     c("pub", "restaurant"),